# services/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Smart Parks Protect",
  "version": "committed"
 },
 "paths": {
...
  "/api/v1/projects/{project_id}/exports": {
   "post": {
    "tags": [
     "exports"
    ],
    "summary": "Create Export",
    "description": "Queue an export job. The export service writes the file to MinIO; poll the job, then\ndownload it. Nothing is limited in size here, the file streams from the database.",
    "operationId": "create_export_api_v1_projects__project_id__exports_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExportParameters"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExportJobRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "List Exports",
    "operationId": "list_exports_api_v1_projects__project_id__exports_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "key of the last item of the previous page",
       "title": "Cursor"
      },
      "description": "key of the last item of the previous page"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PageResponse_ExportJobRead_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/exports/direct": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Direct",
    "description": "Download a small export at once (at most DIRECT_MAX_ROWS rows, architecture 13.8).\nLarger requests get 413 and should become a job.",
    "operationId": "direct_api_v1_projects__project_id__exports_direct_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     },
     {
      "name": "dataset",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/ExportDataset"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": true,
      "schema": {
       "$ref": "#/components/schemas/ExportFormat"
      }
     },
     {
      "name": "time_from",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time",
       "title": "Time From"
      }
     },
     {
      "name": "time_to",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time",
       "title": "Time To"
      }
     },
     {
      "name": "entity_ids",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string",
        "format": "uuid"
       },
       "maxItems": 500,
       "title": "Entity Ids"
      }
     },
     {
      "name": "device_ids",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string",
        "format": "uuid"
       },
       "maxItems": 500,
       "title": "Device Ids"
      }
     },
     {
      "name": "metric_keys",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "maxItems": 100,
       "title": "Metric Keys"
      }
     },
     {
      "name": "data_source_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Data Source Id"
      }
     },
     {
      "name": "timezone",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "IANA name; times are written in this zone",
       "default": "UTC",
       "title": "Timezone"
      },
      "description": "IANA name; times are written in this zone"
     },
     {
      "name": "include_names",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Entity, device and metric names as columns",
       "default": true,
       "title": "Include Names"
      },
      "description": "Entity, device and metric names as columns"
     },
     {
      "name": "bucket",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Ladder key, `all`, or empty for automatic",
       "title": "Bucket"
      },
      "description": "Ladder key, `all`, or empty for automatic"
     },
     {
      "name": "aggregates",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "$ref": "#/components/schemas/Aggregate"
       },
       "title": "Aggregates"
      }
     },
     {
      "name": "group_by",
      "in": "query",
      "required": false,
      "schema": {
       "$ref": "#/components/schemas/GroupBy",
       "default": "entity"
      }
     },
     {
      "name": "layout",
      "in": "query",
      "required": false,
      "schema": {
       "$ref": "#/components/schemas/Layout",
       "default": "long"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/exports/{job_id}": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Get Export",
    "operationId": "get_export_api_v1_projects__project_id__exports__job_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Job Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExportJobRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/exports/{job_id}/download": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Download Export",
    "description": "The finished file, streamed from MinIO through the API (the browser never talks to\nMinIO).",
    "operationId": "download_export_api_v1_projects__project_id__exports__job_id__download_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Job Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/exports/{job_id}/reproduce": {
   "post": {
    "tags": [
     "exports"
    ],
    "summary": "Reproduce Export",
    "description": "Run the same parameters again as a new job (architecture 14: reproducible from explicit\nfilters). Late data makes the result differ; the metadata of both jobs shows how.",
    "operationId": "reproduce_export_api_v1_projects__project_id__exports__job_id__reproduce_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Job Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExportJobRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Aggregate": {
    "type": "string",
    "enum": [
     "mean",
     "min",
     "max",
     "median",
     "sum",
     "count",
     "first",
     "last"
    ],
    "title": "Aggregate"
   },
...
   "ExportDataset": {
    "type": "string",
    "enum": [
     "source_events",
     "positions",
     "measurements",
     "aggregates"
    ],
    "title": "ExportDataset",
    "description": "What an export contains (architecture 14, data level)."
   },
   "ExportFormat": {
    "type": "string",
    "enum": [
     "csv",
     "xlsx",
     "json",
     "geojson",
     "gpx"
    ],
    "title": "ExportFormat"
   },
   "ExportJobRead": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "project_id": {
      "type": "string",
      "format": "uuid",
      "title": "Project Id"
     },
     "created_by": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "dataset": {
      "type": "string",
      "title": "Dataset"
     },
     "format": {
      "type": "string",
      "title": "Format"
     },
     "parameters": {
      "additionalProperties": true,
      "type": "object",
      "title": "Parameters"
     },
     "metadata": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata"
     },
     "source_job_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Job Id"
     },
     "progress_rows": {
      "type": "integer",
      "title": "Progress Rows"
     },
     "row_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Row Count"
     },
     "size_bytes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Size Bytes"
     },
     "sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sha256"
     },
     "error_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Code"
     },
     "error_message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error Message"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "expires_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "project_id",
     "created_by",
     "status",
     "dataset",
     "format",
     "parameters",
     "metadata",
     "source_job_id",
     "progress_rows",
     "row_count",
     "size_bytes",
     "sha256",
     "error_code",
     "error_message",
     "created_at",
     "started_at",
     "finished_at",
     "expires_at"
    ],
    "title": "ExportJobRead"
   },
   "ExportParameters": {
    "properties": {
     "dataset": {
      "$ref": "#/components/schemas/ExportDataset"
     },
     "format": {
      "$ref": "#/components/schemas/ExportFormat"
     },
     "time_from": {
      "type": "string",
      "format": "date-time",
      "title": "Time From"
     },
     "time_to": {
      "type": "string",
      "format": "date-time",
      "title": "Time To"
     },
     "entity_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "maxItems": 500,
      "title": "Entity Ids"
     },
     "device_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "maxItems": 500,
      "title": "Device Ids"
     },
     "metric_keys": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 100,
      "title": "Metric Keys"
     },
     "data_source_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Source Id"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone",
      "description": "IANA name; times are written in this zone",
      "default": "UTC"
     },
     "include_names": {
      "type": "boolean",
      "title": "Include Names",
      "description": "Entity, device and metric names as columns",
      "default": true
     },
     "bucket": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bucket",
      "description": "Ladder key, `all`, or empty for automatic"
     },
     "aggregates": {
      "items": {
       "$ref": "#/components/schemas/Aggregate"
      },
      "type": "array",
      "title": "Aggregates"
     },
     "group_by": {
      "$ref": "#/components/schemas/GroupBy",
      "default": "entity"
     },
     "layout": {
      "$ref": "#/components/schemas/Layout",
      "default": "long"
     }
    },
    "type": "object",
    "required": [
     "dataset",
     "format",
     "time_from",
     "time_to"
    ],
    "title": "ExportParameters",
    "description": "Everything an export needs; stored on the job so it can be reproduced."
   },
...
   "GroupBy": {
    "type": "string",
    "enum": [
     "entity",
     "device"
    ],
    "title": "GroupBy"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Layout": {
    "type": "string",
    "enum": [
     "series",
     "long",
     "wide"
    ],
    "title": "Layout"
   },
...
   "PageResponse_ExportJobRead_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ExportJobRead"
      },
      "type": "array",
      "title": "Items"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items"
    ],
    "title": "PageResponse[ExportJobRead]"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    }
   }
  }
 }
}